Added Mocks for data + Simple API test with a mocked Backend.
