+++ Frontend/src/app/app.module.ts
@@ -12,7 +12,7 @@ import { NoteModalComponent } from "app/components/note-modal/note-modal.compone
 import { PageNotFoundComponent } from "app/views/page-not-found/page-not-found.component";
 import { NavbarComponent } from "app/components/navbar/navbar.component";
 import { ApCompetencesComponent } from "app/components/ap-competences/ap-competences.component";
-import { GlobalVariablesService } from "app/services/global-variables/global-variables.service";
+import { GlobalVariablesService } from "app/services/global-variables.service";
 import { KeysPipe } from "app/pipes/keys/keys.pipe";
 
 const appRoutes: Routes = [
