+++ Frontend/src/app/services/api-mocking-tests/user.mock.ts
@@ -1,0 +1,3 @@
+import { User } from "app/classes/user";
+
+export const MockUser: User = new User('tesb1234', 'Bob', 'Testeur', 'bob.testeur@usherbrooke.ca');
