+++ Frontend/src/app/views/main/main.component.ts
@@ -3,7 +3,7 @@ import { Evaluation } from "app/classes/evaluation";
 import { Ap } from "app/classes/ap";
 import { NoteModalComponent } from "app/components/note-modal/note-modal.component";
 import { Competence } from "app/classes/competence";
-import { GlobalVariablesService } from "app/services/global-variables/global-variables.service";
+import { GlobalVariablesService } from "app/services/global-variables.service";
 
 @Component({
   selector: 'app-main',
